Pages directory separate from components
404 Not Found error page
