+++ src/pages/NotFound.tsx
@@ -1,0 +1,10 @@
+import Link from "../components/common/Link";
+
+const NotFound = () => (
+  <>
+    <h1>Uh oh! There's nothing here!</h1>
+    <Link to="/">Go home</Link>
+  </>
+);
+
+export default NotFound;
