+++ src/routes.tsx
@@ -1,7 +1,8 @@
 import { RouteObject } from "react-router-dom";
-import Landing from "./components/Landing";
-import Social from "./components/Social";
-import Scramble from "./components/Scramble/Scramble";
+import Scramble from "./pages/Scramble/Scramble";
+import Landing from "./pages/Landing";
+import Social from "./pages/Social";
+import NotFound from "./pages/NotFound";
 
 const routes: RouteObject[] = [
   {
@@ -10,12 +11,16 @@ const routes: RouteObject[] = [
   },
   {
     path: "/social",
-    element: <Social />
+    element: <Social />,
   },
   {
     path: "/cube/scramble",
-    element: <Scramble />
-  }
+    element: <Scramble />,
+  },
+  {
+    path: "*",
+    element: <NotFound />,
+  },
 ];
 
 export default routes;
